장착한 Weapon 위젯 구현(무기 이미지, 탄약 수)

+++ Source/Distortion/GameMode/DT_GameMode.cpp
@@ -55,7 +55,7 @@ void ADT_GameMode::RequestPlayerRespawn(ACharacter* Character, AController* Cont
 	{
 		TArray<AActor*> PlayerStarts;
 		UGameplayStatics::GetAllActorsOfClass(this, APlayerStart::StaticClass(), PlayerStarts);
-		int32 Selection = FMath::RandRange(0, PlayerStarts.Num() - 1);
+		uint8 Selection = FMath::RandRange(0, PlayerStarts.Num() - 1);
 		
 		RestartPlayerAtPlayerStart(Controller, PlayerStarts[Selection]);
